requirements.txt
streamlit==1.28.0
confluent-kafka==2.3.0
pandas==2.1.4
plotly==5.15.0
numpy==1.26.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.32.3
apscheduler==3.10.0
